chore(ext/cache): make helper functions public (#16117)
