+++ ext/cache/Cargo.toml
@@ -2,7 +2,7 @@
 
 [package]
 name = "deno_cache"
-version = "0.2.0"
+version = "0.3.0"
 authors = ["the Deno authors"]
 edition = "2021"
 license = "MIT"
